pyth/board_card/pyproject.toml
[tool.poetry]
name = "board_card"
version = "0.1.0"
description = ""
authors = ["bussiere <bussiere@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fpdf = "^1.7.2"
Pillow = "^9.2.0"
natsort = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
